requirements.txt
# ==============================================================================
# MOSAIC: Multi-Omics Survival Analysis with Interpretable Cross-modal Attention
# ==============================================================================

# Core Deep Learning
torch==2.1.0
torchvision==0.16.0
pytorch-lightning==2.1.0
einops==0.7.0
timm==0.9.12

# Hugging Face (for UNI model)
huggingface_hub==0.20.0
transformers>=4.36.0

# Whole Slide Image Processing
openslide-python
histolab
Pillow>=10.0.0

# Data Storage & Processing
h5py==3.10.0
pandas==2.1.0
numpy==1.26.0
scipy==1.11.0

# Single-cell & Omics
scanpy
anndata==0.10.3

# Survival Analysis
lifelines==0.27.8
scikit-survival==0.22.0
pycox==0.2.3

# Model Interpretability
captum==0.6.0
shap==0.44.0

# Visualization
matplotlib==3.8.0
seaborn

# API Serving
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Experiment Tracking
wandb==0.16.0

# Utilities
tqdm==4.66.0
PyYAML>=6.0
scikit-learn>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.26.0
